Put UserProfile thunk actions in actions.thunk file
